# server/openapi.yaml
openapi: 3.0.0
info:
  title: Game Monitor API
  version: 1.0.0
  description: REST API for multiplayer microgames platform
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
...
  /auth/register:
    post:
      summary: Register with email
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        '200':
          description: User registered
        '400':
          description: Invalid input
  /auth/login:
    post:
      summary: Login with email
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
  /users/{id}:
    get:
      summary: Get user profile
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
        '404':
          description: User not found
...
  /matches:
    get:
      summary: Get user match history
      tags: [Matches]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Match history
  /matches/{matchId}:
    get:
      summary: Get match details
      tags: [Matches]
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Match details
...